desktop-tool/requirements.txt
pyaudio==0.2.11
speech-recognition==3.10.0
pynput==1.7.6
requests==2.31.0
python-dotenv==1.0.0
keyboard==0.13.5

# 语音识别提供商 SDK（可选安装）
# 本地 Whisper
openai-whisper==20231117

# OpenAI API
openai==1.3.0

# 腾讯云 SDK
tencentcloud-sdk-python==3.0.1056

# 阿里云 SDK
alibabacloud_nls_meta20190103==1.0.0

# 百度云 SDK
baidu-aip==4.16.10
